+++ gulpfile.js
@@ -143,7 +143,15 @@ gulp.task("build-version-wordpress-plugin", function() {
 			path: [`./wordpress/`],
 			inheritExtension: true,
 		}))
-		.pipe(gulp.dest(`./dist/wordpress/`));
+		.pipe(gulp.dest(`./dist/wordpress/last-9-photos-webcomponent/`));
+});
+
+gulp.task("cpy-js-wordpress-plugin", function() {
+	let files = [
+		"./dist/lib/**/*"
+	];
+
+	return gulp.src(files).pipe(gulp.dest(`./dist/wordpress/last-9-photos-webcomponent/js/`));
 });
 
 /**
@@ -186,4 +194,4 @@ gulp.task("browser-sync", function() {
 gulp.task("start", gulp.series("build-css", "build-html", "build-js", "browser-sync"));
 gulp.task("dist", gulp.series("build-css", "build-html", "build-js", "build-js-minify"));
 gulp.task("release", gulp.series("build-version-latest-lib", "build-version-latest-lib-minify"));
-gulp.task("wordpress-plugin", gulp.series("build-version-wordpress-plugin"));
+gulp.task("wordpress-plugin", gulp.series("build-version-wordpress-plugin", "cpy-js-wordpress-plugin"));