+++ webcomponent/js/main.js
@@ -13,7 +13,7 @@ class InstagramWidget extends HTMLElement {
 		super();
 
 		const template = document.createElement("template");
-		template.innerHTML = `<style id="instagram-widget-style">{% include 'css/main.css' %}</style>{% include 'main.html' %}`;
+		template.innerHTML = `<style id="instagram-widget-style" part="style">{% include 'css/main.css' %}</style>{% include 'main.html' %}`;
 
 		this.attachShadow({mode: "open"});
 		this.shadowRoot.appendChild(template.content.cloneNode(true));
@@ -27,12 +27,14 @@ class InstagramWidget extends HTMLElement {
 			"cache": "enabled",
 			"border-spacing": "2px",
 			"border-corners": "5",
-			"force-square": "yes"
+			"force-square": "yes",
+			"shadows": "disabled",
+			"mouse-hover": "disabled"
 		};
 
 		this.options = Object.create(this.options_default);
 
-		this.resize_event = function(event) {
+		this.resizeAction = function(event) {
 			this.resize(event);
 		}.bind(this);
 	}
@@ -43,7 +45,7 @@ class InstagramWidget extends HTMLElement {
 	 *
 	 */
 	connectedCallback() {
-		window.addEventListener("resize", this.resize_event);
+		window.addEventListener("resize", this.resizeAction);
 	}
 
 	/**
@@ -51,7 +53,7 @@ class InstagramWidget extends HTMLElement {
 	 * =====================
 	 *
 	 */
-	build_html() {
+	buildHTML() {
 		let data = this.json.graphql.user.edge_owner_to_timeline_media.edges;
 
 		let photos = [];
@@ -68,14 +70,55 @@ class InstagramWidget extends HTMLElement {
 
 		let html = "";
 		for (let i = 0; i < photos.length && i < this.options["items-limit"]; i++) {
-			html += `<li><a href="${photos[i].url}" rel="nofollow external noopener noreferrer" target="_blank" title="${photos[i].caption.substring(0, 100).replace(/"/g, "")}"><img width="${this.options["image-width"]}" height="${this.options["image-height"]}" src="${photos[i].display_url}" alt="${photos[i].caption.substring(0, 100).replace(/"/g, "")}" loading="lazy" /></a></li>`;
+			html += `<li class="instagram-widget-li" part="li li-${i}"><a href="${photos[i].url}" rel="nofollow external noopener noreferrer" target="_blank" title="${photos[i].caption.substring(0, 100).replace(/"/g, "")}" class="instagram-widget-link" part="link link-${i}"><img width="${this.options["image-width"]}" height="${this.options["image-height"]}" src="${photos[i].display_url}" alt="${photos[i].caption.substring(0, 100).replace(/"/g, "")}" loading="lazy" class="instagram-widget-photo" part="photo photo-${i}" /></a></li>`;
 		}
 		this.shadowRoot.querySelector(".instagram-widget-photos").innerHTML = html;
+
+		switch (this.options["mouse-hover"]) {
+			case "type1":
+				this.shadowRoot.querySelector("#instagram-widget-style").innerHTML = `${this.shadowRoot.querySelector("#instagram-widget-style").innerHTML}
+				#instagram-widget .instagram-widget-li{overflow: hidden;}
+				#instagram-widget .instagram-widget-photo{transition: transform .5s ease;}
+				#instagram-widget .instagram-widget-li .instagram-widget-photo:hover{transform: scale(1.5);}`;
+				break;
+
+			case "type2":
+				this.shadowRoot.querySelector("#instagram-widget-style").innerHTML = `${this.shadowRoot.querySelector("#instagram-widget-style").innerHTML}
+				#instagram-widget .instagram-widget-li{overflow: hidden;}
+				#instagram-widget .instagram-widget-photo{transition: transform .5s ease;}
+				#instagram-widget .instagram-widget-li:hover .instagram-widget-photo{transform: scale(2) rotate(25deg);}`;
+				break;
+
+			case "type3":
+				this.shadowRoot.querySelector("#instagram-widget-style").innerHTML = `${this.shadowRoot.querySelector("#instagram-widget-style").innerHTML}
+				#instagram-widget .instagram-widget-li{overflow: hidden;}
+				#instagram-widget .instagram-widget-photo{transition: transform 2s, filter 1.5s ease-in-out;transform-origin: center center;filter: brightness(50%);}
+				#instagram-widget .instagram-widget-li:hover .instagram-widget-photo{filter: brightness(100%);transform: scale(1.3);}`;
+				break;
+
+			case "type4":
+				this.shadowRoot.querySelector("#instagram-widget-style").innerHTML = `${this.shadowRoot.querySelector("#instagram-widget-style").innerHTML}
+				#instagram-widget .instagram-widget-li{overflow: hidden;}
+				#instagram-widget .instagram-widget-photo{transition: transform 1s, filter 2s ease-in-out;filter: blur(2px);transform: scale(1.2);}
+				#instagram-widget .instagram-widget-li:hover .instagram-widget-photo{filter: blur(0);transform: scale(1);}`;
+				break;
+
+			case "type5":
+				this.shadowRoot.querySelector("#instagram-widget-style").innerHTML = `${this.shadowRoot.querySelector("#instagram-widget-style").innerHTML}
+				#instagram-widget .instagram-widget-li{overflow: hidden;}
+				#instagram-widget .instagram-widget-photo{transition: transform .5s ease-in-out;transform: scale(1.4);transform-origin: 0 0;}
+				#instagram-widget .instagram-widget-li:hover .instagram-widget-photo{transform: scale(1.25) translateY(-30%);}`;
+				break;
+
+			default:
+				break;
+		}
 
 		if (this.options["grid"] !== "" && this.options["grid"] !== null && this.options["grid"] !== "responsive") {
 			let grid = this.options["grid"].split("x");
 			let width = 100 / parseInt(grid[0]);
 			let images = this.shadowRoot.querySelectorAll(".instagram-widget-photos img");
+			let li = this.shadowRoot.querySelectorAll(".instagram-widget-li");
 			for (let i=0; i < images.length; i++) {
 				images[i].removeAttribute("width");
 				images[i].style.width = `calc(${(width)}% - (${this.options["border-spacing"]} * (${parseInt(grid[0])} * 2)))`;
@@ -84,22 +127,61 @@ class InstagramWidget extends HTMLElement {
 				images[i].style.borderRadius = `${this.options["border-corners"]}%`;
 				images[i].style.margin = this.options["border-spacing"];
 
+				switch (this.options["shadows"]) {
+					case "type1":
+						images[i].style.boxShadow = "0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24)";
+						break;
+
+					case "type2":
+						images[i].style.boxShadow = "0 14px 28px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22)";
+						break;
+
+					case "type3":
+						images[i].style.boxShadow = "0 19px 38px rgba(0,0,0,0.30), 0 15px 12px rgba(0,0,0,0.22)";
+						break;
+
+					default:
+						break;
+				}
+
 				if (this.options["force-square"] === "yes") {
 					images[i].removeAttribute("height");
 					images[i].style.height = `${this.shadowRoot.querySelector(".instagram-widget-photos img").clientWidth}px`;
 				}
+
+				li[i].style.maxHeight = `${this.shadowRoot.querySelector(".instagram-widget-photos img").clientHeight}px`;
 			}
 		} else {
+			let li = this.shadowRoot.querySelectorAll(".instagram-widget-li");
 			let images = this.shadowRoot.querySelectorAll(".instagram-widget-photos img");
 			for (let i=0; i < images.length; i++) {
 				images[i].style.borderRadius = `${this.options["border-corners"]}%`;
 				images[i].style.margin = this.options["border-spacing"];
+
+				switch (this.options["shadows"]) {
+					case "type1":
+						images[i].style.boxShadow = "0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24)";
+						break;
+
+					case "type2":
+						images[i].style.boxShadow = "0 14px 28px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22)";
+						break;
+
+					case "type3":
+						images[i].style.boxShadow = "0 19px 38px rgba(0,0,0,0.30), 0 15px 12px rgba(0,0,0,0.22)";
+						break;
+
+					default:
+						break;
+				}
 
 				if (this.options["force-square"] === "yes") {
 					images[i].removeAttribute("height");
 					images[i].style.maxHeight = "none";
 					images[i].style.height = `${this.shadowRoot.querySelector(".instagram-widget-photos img").clientWidth}px`;
 				}
+
+				li[i].style.maxHeight = `${this.shadowRoot.querySelector(".instagram-widget-photos img").clientHeight}px`;
 			}
 		}
 	}
@@ -123,7 +205,7 @@ class InstagramWidget extends HTMLElement {
 	 * =====================
 	 *
 	 */
-	api_fetch() {
+	apiFetch() {
 		let url = `https://www.instagram.com/${this.options["username"].replace("@", "")}/?__a=1`;
 		fetch(url, {"cache": this.options["cache"] === null || this.options["cache"] === "enabled" ? "force-cache" : "default"}).then(function(response) {
 			if (response.status === 200) {
@@ -131,12 +213,12 @@ class InstagramWidget extends HTMLElement {
 			}
 		}).then(function(response) {
 			this.json = response;
-			this.build_html();
+			this.buildHTML();
 		}.bind(this));
 	}
 
 	static get observedAttributes() {
-		return ["username", "items-limit", "grid", "image-width", "image-height", "border-spacing", "border-corners", "force-square", "cache"];
+		return ["username", "items-limit", "grid", "image-width", "image-height", "border-spacing", "border-corners", "force-square", "mouse-hover", "shadows", "cache"];
 	}
 
 	attributeChangedCallback(name_attribute, old_vale, new_value) {
@@ -149,11 +231,11 @@ class InstagramWidget extends HTMLElement {
 
 			switch (name_attribute) {
 				case "username":
-				  this.api_fetch();
+				  this.apiFetch();
 				  break;
 				default:
 				  if (this.json !== null) {
-						this.build_html();
+						this.buildHTML();
 				  }
 			  }
 		}
@@ -165,7 +247,7 @@ class InstagramWidget extends HTMLElement {
 	 *
 	 */
 	disconnectedCallback() {
-		window.removeEventListener("resize", this.resize_event);
+		window.removeEventListener("resize", this.resizeAction);
 	}
 }