+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-# v2.6.0 (TBD)
-* Wordpress Plugin: Integration with Gutemberg
+# v2.6.0 (May 17, 2020)
+* Fix: Wordpress Plugin
+* New attribute: mouse-hover
+* New attribute: shadows
+* Feature: Overwrite CSS Style with selector `::part`
 
 [![](https://img.shields.io/badge/donate-paypal-005EA6.svg?logo=paypal)](https://www.paypal.me/ptkdev) [![](https://img.shields.io/badge/donate-patreon-F87668.svg?logo=patreon)](https://www.patreon.com/ptkdev) [![](https://img.shields.io/badge/donate-sponsors-ea4aaa.svg?logo=github)](https://github.com/sponsors/ptkdev/)  [![](https://img.shields.io/badge/donate-ko--fi-29abe0.svg?logo=ko-fi)](https://ko-fi.com/ptkdev)